requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
pdf2image>=3.1.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0